SC_Project_1718-ARS/backend/machine_learning/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
CacheControl==0.12.3
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
Django==1.11.5
djangorestframework==3.6.4
idna==2.6
imdbpie==4.3.0
lockfile==0.12.2
matplotlib==2.0.2
msgpack-python==0.4.8
numpy==1.13.1
panda==0.3.1
pandas==0.20.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.9.dev0
selenium==3.6.0
six==1.11.0
sklearn==0.0
urllib3==1.22
